requirements.txt
numpy==1.20.0
pandas==1.4.2
PyYAML==6.0
scikit_learn==1.1.3
scipy==1.8.0
torch==2.0.0+cu118
tqdm==4.64.0
